backend/requirements.txt
# Core API Framework
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1

# Machine Learning - Railway Track
scikit-learn==1.5.2
xgboost==2.1.3
joblib==1.4.2

# Machine Learning - Metro APU (Deep Learning)
tensorflow==2.20.0
numpy==1.26.4

# LLM Integration
openai==1.57.4
anthropic==0.42.0

# CORS
python-multipart==0.0.20
